Extract AbstractWebEngineContext
